enhance: move StreamingNodeCataLog interface into shared pkg/v3 metastore
streamingnode has no MetaTable facade; its persistent meta lives in the WAL
recovery layer and only consumes StreamingNodeCataLog. To let the shared
module own the contract (so the pooled catalog service can depend on it),
move the StreamingNodeCataLog interface from internal/metastore into
pkg/v3/metastore, and re-export it via a type alias from the internal package
so existing importers compile unchanged. The streamingnode KV implementation
stays in internal, mirroring the rootcoord etcd impl.

related: #50917
